Synchronise between the deletion operation and the watch subsystem.
Enable the oBIX server delete a subnode from a device contract as long
as the impact subnode is writable.

As a result any watch meta tags in the deleted subtree will also be
removed, which requires that corresponding watch_item_t descriptor
must be marked as invalidated so as not to corrupt the watch subsystem.

Also any pending poll tasks should be activated so that a null object
can be returned back to clients immediately.

BTW, clients should manually remove an invalidatd watch item so that
they can have the same href monitored once again after they get to know
that the owner of the device contract has installed a new node on that
href.

Last but not least, the signOff request would have to be used in order
to remove an entire device contract. The signOff handler will mainpulate
a new device subsystem to achieve certain access control on device
contracts. But it has not been implemented yet.