Enable adding nodes to non-writable objects.
Previously adding or deleting nodes from the given node all require that
the target node itself or its parent must be writable.

However, if a child of a non-writable object is mistakenly deleted, there
is no way to re-install it because its parent is not writable.

To mitigate such situation, adding child nodes won't require the target
node to be writable, but changing its val attribute of deleting itself
or any of its child node still does.

However, the added node should better be writable, otherwise it cannot
be delete if its parent is also not writable.

The problem can be fixed properly once the signOff handler is implemented.
If a child node of a non-writable object is inadvertantly deleted, simply
sign off the entire parent object and re-signUp.

Signed-off-by: Qingtao Cao <harry.cao@nextdc.com>