Remove XPath usage from history initialisation code.
As a matter of fact, the index file of a history facility is very
straightforward which contains just a list of file abstract objects
for all existing raw history data files, therefore a simple for
loop suits best to traverse all items in the list. Whereas XPath
will be an overkill in this case.