Avoid unconditional copy of node in the PUT handler
The xmldb_update_node used to unconditionally copy the input root node
before reparenting it to the global DOM tree to avoid corruption of
the global DOM tree. Now it is made clear that this relates to the XML
parser's internal dictionary and an extra copy of the input root node
will only take place if a dictionary has been used.

To start with, the global DOM tree is made independent from any parser's
internal dictionary. Optimisation could be done in the near future once
we figure out how to use XML parser dictionary in a thread-safe manner.