Minor improvement on the signUp handler.
The "writable" attribute is removed immediately at the entrance of
the signUp handler so that it won't be set in a device contract
in the global DOM tree nor would it appear in the response of the
signUp request.

Also xmldb_copy_node should be preferred to xml_copy when it comes
to node in the global DOM tree, since the former will call
xmlUnlinkNode to de-associate the copied node from the global DOM
tree properly.