Support ISO-8601 timestamp format
Support the timezone designator in the timestamp string according to
the ISO-8601 standard.

If no timezone designator is available, then it will be regarded as in
UTC timezone. Other valid format includes the following:
 . "Z"
 . "+/-hh", where "hh" must be 2 digits no larger than 12;
 . "+/-hhmm", where "hh" must be 2 digits no larger than 12 and
   in theory "mm" must be no larger than 59. In particular, if "hh"
   equals to 12 then "mm" must be "00"

It's worthwhile to note that the current strptime() API implementation
has adopted practical compromise from the timezone definition as in
ISO-8601, for example:
 . no colon supported. At most 4 digits will be consumed after the "+/-"
   offset character. Moreover, if it encounters any non-digit character,
   it stops parsing the rest of timezone string;
 . the "mm" value will not always be treated as "minutes" as expected
   by user. In theory, it must be a multiple of 3. In practice, it should
   be a multiple of 15.

Please refer to docs/timezone.md file for details.

Also included is a test case to demonstrate all the tricks on timestamp
strings and relevant test results.