Fix the start boundary of a history record
"<obj " used to be regarded as the start of a history record. However,
unfortunately, with the presence of complex device contract and relevant
history record that contains sub nodes, this will simply break in that
parse_log() will mistakenly regard the beginning of a sub node in the
previous record of the desirable one as the starting point of it, resulting
segment of previous history record fetched as part of response as well.

To fix it, simply be stringent on the starting point of a history record
to be "<obj is=\"obix:HistoryRecord\">"

Signed-off-by: Qingtao Cao <harry.cao@nextdc.com>