Fix a segfault when signUp through the batch mechanism - Issue#91
The signUp handler is the only exception from all the other handlers
which return a copy of some node in the global DOM tree. Such copy
will have no connection with any XML parser dictionary at all and
they can be safely released along with another temporary document
created for response purpose.

When the signUp handler is directly invoked, the node from the input
document which manipulates a XML parser dictionary is directly returned.
In order not to interfere with the dictionary, obix_server_reply_object
will reference it properly.

Unfortunately, this won't work if the signUp handler is invoked via
the batch mechanism, because the input node is not added as the root
node of the temporary document but as children of the batchOut contract,
consequently obix_server_reply_object could not identify the usage
of a dictionary.

To address this, an extra copy is made before re-parenting the returned
node from a handler into the batchOut contract. If it has an owner
document which further takes advantage of a dictionary, it will be copied
so as to de-associate from the dictionary properly.

Signed-off-by: Qingtao Cao <harry.cao@nextdc.com>