Enable handlers to be redirected to via the batch mechanism
Also fixing Issue#94 and #98

If handlers are directly invoked through the GET/PUT/POST methods,
the requested URI is recorded in the obix_request_t structure. However,
if handlers are invoked via the batch mechanism, the requested URI is
always "/obix/batch" and the URI for the specific facility needed is
is specified by the "val" attribute of a batch command in the batchIn
contract.

To this end, handlers should take an extra "overrideUri" parameter to
manipulate the redirected-to URI in the batch request.

It's worthwhile to mention that not all handlers can be invoked via
the batch facility, for example, watch.pollChanges handler and history
handlers compete against the current server thread handling the batchIn
request on sending responses through the same FCGI request independently.
Worse still, the watch.pollChanges handler will race to release the FCGI
request resulting in segfault when the current server thread tries to
send back batchOut contract. Therefore they should not be allowed to be
raised through the batch mechanism.

Last but not least, remove redundant function calls in batch.c and perform
sanity checks on the redirected-to URI to prevent malicious access to the
root node of the global DOM tree.