New Device subsystem
Introduce a new Device subsystem to manage the registration and
unregistration of device contracts to and from the global XML DOM tree.

Basically a descriptor is setup for a newly registered device and added
to the hash table of the Device subsystem. Each time a href is accessed,
the hash table (and its auxiliary cache facility) is consulted first to
get the target device's descriptor, which hosts a number of data
structures such as its owner's ID, the pointer to its node in the XML
DOM tree, and the tsync_t facility to ensure thread-safe access etc.

Each time a device is read or written, relevant attempts are placed into
the "read region" and "write region" of the device respectively so that
device contracts can be concurrently accessed by normal readers / writers
and polling threads which are all asynchronous to others.

With the help of device descriptors, getting the absolute href of a given
node in the XML DOM tree has become really fast. Unlike the old method
used in the xmldb_node_path() to traverse from the given node all the way
back to the top (which is a performance-killer and non thread-safe), the
search operation now only happens within the boundary of relevant device
contract.

Also implemented is the signOff handler and relevant API on the client side.
oBIX clients now can unregister the devices before exit, which also helps
simplify the logic in the write handler which should only concern about
changing the val attribute of the given node instead of adding or deleting
sub nodes. All such requests must go through the signUp and signOff handlers
respectively for sake of access control and device management.

Furthermore, with the help of device descriptors the "Persistent Devices"
feature is a low-hanging fruit to pick up. Device persistent files on the
hard drive are created and deleted at signUp and signOff respectively, and
the device contract is saved onto the hard drive when needed. If the oBIX
server crashed and re-started, it would re-start from existing persistent
files and re-create relevant device structures so as to resume processing
relevant requests from client applications which don't care about and aren't
involved with oBIX server recovery at all.

BTW, the current adaptors used in production (the BMS adaptor and the MGATE
adaptor) have not utilised the watch objects at all. For new adaptors to be
developed in the future that manipulates watch objects, a similar feature on
the oBIX server, the "Persistent Watches", should be developed in order to
restore watches used by clients.

In the future a new Security Engine can be further developed to adopt at
least DAC and hopefully MAC to the oBIX server. The binary of user-specific
access rules database can be loaded into the oBIX server at start-up, which
is consulted by relevant hooks in the Device / History / Watch subsystems
before any operation is carried out.

Last but not least, please refer to the technical deep dive document in
docs/ for insights about the implementation of the Device subsystem