Increase robustness of history facility when no disk space
No O_TRUNC option should be applied when opening a history index file.

The write operation to the index file may fail due to lack of disk
space, so they should not be truncated to zero at the time of open
otherwise when write fails there will be erased index files left
on the hard drive breaching the history facility.

Also manipulate the global errno variable to tell if the creation of
history facility on the hard drive is failed due to no available space.

Lastly, enrich the HISTORY.md file with steps to setup relevant
testing environment.

Signed-off-by: Qingtao Cao <qingtao.cao.au@gmail.com>