Revert "Have mg_adaptor skip over off-lined devices."
This reverts commit 7f29bbd9f4b3826307063ff0cdb66c84c8688ecf.

IT admins would favour the idea not having to have them send USR1
signals to mark a device on-lined once it gets re-installed properly.
To this end the previous behaviour of the mg_adaptor is preserved
except that no error logs about "Connection timed out" for off-lined
devices are generated.