Disable signal usage in libcurl to avoid a critical defect
There is a well-known defect in libcurl when using signals to timeout
during DNS lookups resulting in oBIX application unexpected termination
on RHEL6 (but not on Fedora).

The ideal solution is to upgrade libcurl or re-compile it with the c-ares
package which implements async DNS lookup instead of using SIGALM signal.
However, it is discouraged by IT team but to use config option to indicate
whether signals should be disabled in applications configuration file so
that they can tune it accordingly.