Fix a defect in xml_write_file - Issue #118
Now that O_TRUNC flag no longer adopted when a XML file is opened
for update, the size of the latest write attempt must be explicitly
set as the new length of the file. Otherwise the tail of previous
content will be preserved if the new update written less amount
of data resulting in mal-formed XML file!