Improve comments about listen-backlog parameter
Improve the comments about the listen-backlog parameter, which is
passed to the listen() system call directly to decide the depth of
the backlog queue of the FCGX listening socket on the oBIX server.

It's high time we developed the multi-thread support of the oBIX
server so as to consume established requests from the backlog queue
quickly enough to avoid the client side getting HTTP 503 "Service
Not Available" type of error.

Last but not least, the backlog queue depth is increased to mitigate
the situation until the multi-thread support has been done. Once the
multi-thread has been enabled, its depth won't make much difference.